pret

+++ sql/conception.sql
@@ -2,7 +2,7 @@
 CREATE TABLE finance_Etablissement_Financier (
    id INTEGER PRIMARY KEY AUTO_INCREMENT,
    nom_etablissement VARCHAR(50),
-   fonds INTEGER
+   fonds DECIMAL(10,2)
 );
 
 CREATE TABLE finance_Type_Pret (
@@ -23,11 +23,11 @@ CREATE TABLE finance_Clients (
 CREATE TABLE finance_Pret (
    id_type_pret INTEGER,
    id INTEGER PRIMARY KEY AUTO_INCREMENT,
-   montant_mensuel INTEGER,
-   montant_pret INTEGER,
+   montant_mensuel DECIMAL(10,2),
+   montant_pret DECIMAL(10,2),
    date_pret DATE,
    date_fin DATE,
-   assurance INTEGER,
+   assurance DECIMAL(10,2),
    delai INTEGER,
    FOREIGN KEY(id_type_pret) REFERENCES finance_Type_Pret(id)
 );
@@ -40,8 +40,8 @@ CREATE TABLE finance_Type_Transaction (
 CREATE TABLE finance_Historique_Pret (
    id_pret INTEGER,
    id INTEGER PRIMARY KEY AUTO_INCREMENT,
-   capital INTEGER,
-   interet INTEGER,
+   capital DECIMAL(10,2),
+   interet DECIMAL(10,2),
    date_paiement DATE,
    FOREIGN KEY(id_pret) REFERENCES finance_Pret(id)
 );
@@ -52,7 +52,7 @@ CREATE TABLE finance_Transaction (
    id_client INTEGER,
    id_type_transaction INTEGER,
    date_transaction DATE,
-   montant INTEGER,
+   montant DECIMAL(10,2),
    FOREIGN KEY(id_ef) REFERENCES finance_Etablissement_Financier(id),
    FOREIGN KEY(id_client) REFERENCES finance_Clients(id),
    FOREIGN KEY(id_type_transaction) REFERENCES finance_Type_Transaction(id)
